# Improve performance of enabling/disabling extensions