--- a/web/src/extensions/ExtensionCard.test.tsx
+++ b/web/src/extensions/ExtensionCard.test.tsx
@@ -32,6 +32,7 @@ describe('ExtensionCard', () => {
                             subject={{ id: 'u', viewerCanAdminister: false }}
                             settingsCascade={{ final: null, subjects: null }}
                             platformContext={NOOP_PLATFORM_CONTEXT}
+                            enabled={false}
                         />
                     </MemoryRouter>
                 )